requirements.txt
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
mysql-connector-python==9.5.0
python-dotenv==1.2.1
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
lxml==6.0.2
mysql-connector-python==9.5.0
python-dotenv==1.2.1
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
